Add Python script to insert data into MySQL database
Added a Python script that connects to a local MySQL database and inserts data into the "Records" table in the "YFMI" database. The script utilizes the `mysql-connector-python` library to establish the connection and execute the SQL query. It supports insertion of sample data including fields for `SAMPLE`, `DT1`, `DT2`, `AT_mV`, `PH_mV`, and `TDS_mV`.